Improve the translation mechanism
Adds the LanguageTexts and LanguageAbbrs enums in order to get
more abstraction and organization in the code.

Adds the LANGUAGE_TEXTS_LIST and LANGUAGE_ABBRS_LIST constants
that contain the list of language texts and text abbreviations respectively.

Adds the getPathWithLangExtension function that returns
the correct path to a markdown file taking into account the language.
The sceiba-ng project has an example of usage.